+++ libs/shared/src/lib/components/table/cells/ProgressCell.tsx
@@ -10,7 +10,11 @@ type ProgressCellProps = {
 export const ProgressCell = ({ percentWidth }: ProgressCellProps): JSX.Element => {
   return (
     <StyledProgressBarContainer>
-      <StyledActualProgress borderColor="#40D38F" width={percentWidth} color="#D9F6E9" />
+      <StyledActualProgress
+        borderColor="#40D38F"
+        width={percentWidth}
+        color="#D9F6E9"
+      ></StyledActualProgress>
       <StyledProgressNumber>{`${Math.round(percentWidth)}%`}</StyledProgressNumber>
     </StyledProgressBarContainer>
   );
