+++ libs/shared/src/lib/types/index.ts
@@ -2,4 +2,3 @@ export { type MaterialStatus } from './materialStatus';
 export * from './pcsFollowUpStatus';
 export * from './pcsStatus';
 export { type PackageStatus } from './packageStatus';
-export * from './fieldSettings';
