+++ libs/workorderapp/src/lib/config/gardenConfig.ts
@@ -1,17 +1,30 @@
 import { GardenConfig } from '@equinor/workspace-fusion/garden';
 import { WorkOrder } from '..';
+import { ExtendedGardenFields } from '../types';
 import { GardenHeader, GardenItem } from '../ui-garden';
-import { getItemWidth } from '../utils-garden';
+import { getHighlightedColumn, getItemWidth } from '../utils-garden';
 import { fieldSettings } from '../utils-garden/fieldSettings';
+import { sortPackages } from '../utils-garden/sortPackages';
 
-export const gardenConfig: GardenConfig<WorkOrder> = {
+export const gardenConfig: GardenConfig<WorkOrder, ExtendedGardenFields> = {
   getDisplayName: (item) => item.workOrderNumber,
   initialGrouping: { horizontalGroupingAccessor: 'fwp', verticalGroupingKeys: [] },
+  fieldSettings: fieldSettings,
   customViews: {
     customItemView: GardenItem,
     customHeaderView: GardenHeader,
   },
   visuals: {
     calculateItemWidth: getItemWidth,
+    highlightHorizontalColumn: getHighlightedColumn,
+    rowHeight: 30,
+  },
+  intercepters: {
+    postGroupSorting: (data, keys) => {
+      data.forEach(({ items }) => {
+        items = sortPackages(items, ...keys);
+      });
+      return data;
+    },
   },
 };
