+++ libs/workorderapp/src/lib/utils-garden/index.ts
@@ -2,3 +2,4 @@ export { fieldSettings } from './fieldSettings';
 export { getWorkOrderStatuses } from './getWorkOrderStatuses';
 export { getSubtitleHeader } from './getSubtitleHeader';
 export { getItemWidth } from './getItemWidth';
+export { getHighlightedColumn } from './getHighlightedColumn';
