+++ libs/workorderapp/src/lib/utils-garden/sortPackages.ts
@@ -1,9 +1,6 @@
+import { proCoSysStatusPriorityMap } from '@cc-components/shared';
 import { ExtendedGardenFields, WorkOrder } from '../types';
-import {
-  followUpStatusPriorityMap,
-  getStatusFn,
-  proCoSysStatusPriorityMap,
-} from '../utils-statuses';
+import { followUpStatusPriorityMap, getStatusFn } from '../utils-statuses';
 import { getGroupBy } from './getGroupBy';
 
 const getPriorityMap = (groupBy: keyof WorkOrder | ExtendedGardenFields) => {
@@ -14,9 +11,12 @@ const getPriorityMap = (groupBy: keyof WorkOrder | ExtendedGardenFields) => {
  * @param data List of workorders
  * @param groupByKeys List of grouping keys (only uses first item now, which is gardenKey)
  */
-export const sortPackages = (data: WorkOrder[], ...groupByKeys: (keyof WorkOrder)[]) => {
+export const sortPackages = (
+  data: WorkOrder[],
+  ...groupByKeys: (keyof WorkOrder | ExtendedGardenFields)[]
+): WorkOrder[] => {
   //TODO Handle multiple groupByKeys?
-  const groupBy: keyof WorkOrder | ExtendedGardenFields = groupByKeys[0] || 'hwp';
+  const groupBy = groupByKeys[0] || 'hwp';
   const status = getStatusFn(groupBy);
   const priorityMap = getPriorityMap(groupBy);
   const sortBy =
