+++ libs/workorderapp/src/lib/utils-statuses/index.ts
@@ -3,7 +3,6 @@ export {
   followUpStatusPriorityMap,
   followUpColorMap,
 } from './followup';
-export { proCoSysStatusPriorityMap } from './proCoSysStatusPriorityMap';
 export { getStatusFn } from './getStatusFn';
 export { getMatStatusColorByStatus, getMatStatus } from './material';
 export { getMccrStatusColorByStatus } from './getMccrStatusColorByStatus';
